+++ changelog.md
@@ -1,5 +1,9 @@
 # AntelopeQL changelog
 
+### Patch
+
+- Renamed some of the old EOSIO descriptions to Antelope.
+
 ## 1.0.0
 
 - Initial release